requirements.txt
asgiref==3.2.9
astroid==2.4.1
autopep8==1.5.3
bc==0.0.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
boto3==1.14.22
botocore==1.17.22
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
Django==3.0.7
django-cors-headers==3.4.0
docutils==0.15.2
idna==2.10
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.1
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
uuid==1.30
webdriver-manager==3.2.1
wrapt==1.12.1
